requirements.txt
numpy
opencv-python
pillow
onnxruntime
pyyaml
ensemble-boxes
